feat: Fetch alerts with open status only (#167)
The existing call returns the first `count` alerts. Some of these alerts
may be fixed or dismissed, while there may be older open alerts which
were not returned. Moving the filter operation to the backend resolves
this problem.

N.B. Alert `state` is one of `auto_dismissed`, `dismissed`, `fixed`, and
`open`. Of these, we are only interested in `open`